Use seafdav from github
Use our fork of seafdav.

We install the normal seafdav with the correct packaging and then
overwrite it with the git repo. We could just generate the package with
our fork, but that maakes development much harder, since one needs to be
generate a new package for each change and we will need a pipeline for that.

For now lets just use this far simpler method.